Change how configuration and startup is handled
Force the user to set the password in a config file before the container runs "for real". This forces them to use a
config volume, which avoids the situation where they might accidentally invalidate their license.

I also added some verbiage about how it's a bit unclear how the licensing works. On the assumption that it's tied to the
MAC address of the internet-facing NIC, I'm storing that in case it's needed later.

FWIW, I have emailed their support staff to get some clarifications. For instance they say that the Lite version
prohibits it running in a VM, but I don't know if that applies to containers too.